+++ src/select/mod.rs
@@ -1,4 +1,5 @@
 pub mod base;
 pub mod list_item;
 pub mod section;
+pub mod selector;
 mod util;
